+++ basestation/src/anki/cozmo/basestation/cannedAnimationContainer.cpp
@@ -105,7 +105,7 @@ namespace Cozmo {
       PRINT_NAMED_WARNING("CannedAnimationContainer.DefineFromJson.TooManyAnims",
                           "Expecting only one animation per json file, found %lu. "
                           "Will use first: %s",
-                          animationNames.size(), animationNames[0].c_str());
+                          (unsigned long)animationNames.size(), animationNames[0].c_str());
     }
     
     animationName = animationNames[0];
