+++ basestation/src/anki/cozmo/basestation/robot.cpp
@@ -305,7 +305,7 @@ namespace Anki {
                                          "LocalizedTo: <nothing>");
       GetContext()->GetVizManager()->SetText(VizManager::WORLD_ORIGIN, NamedColors::YELLOW,
                                          "WorldOrigin[%lu]: %s",
-                                         _poseOrigins.size(),
+                                         (unsigned long)_poseOrigins.size(),
                                          _worldOrigin->GetName().c_str());
       
       // create a new memory map for this origin
@@ -358,7 +358,7 @@ namespace Anki {
                                          ObjectTypeToString(object->GetType()), _localizedToID.GetValue());
       GetContext()->GetVizManager()->SetText(VizManager::WORLD_ORIGIN, NamedColors::YELLOW,
                                          "WorldOrigin[%lu]: %s",
-                                         _poseOrigins.size(),
+                                         (unsigned long)_poseOrigins.size(),
                                          _worldOrigin->GetName().c_str());
       
       return RESULT_OK;
@@ -2947,7 +2947,7 @@ namespace Anki {
         
         // Check if there is still space in the message
         if (msg.factory_id.size() >= (int)ActiveObjectConstants::MAX_NUM_ACTIVE_OBJECTS) {
-          PRINT_NAMED_WARNING("Robot.ConnectToBlocks.ArrayFull", "Too many objects specified (limit: %lu)", factory_ids.size());
+          PRINT_NAMED_WARNING("Robot.ConnectToBlocks.ArrayFull", "Too many objects specified (limit: %lu)", (unsigned long)factory_ids.size());
           return RESULT_FAIL;
         }
         
@@ -2967,7 +2967,7 @@ namespace Anki {
         }
 
         msg.factory_id.push_back(fid);
-        PRINT_NAMED_INFO("Robot.ConnectToBlocks.FactoryID", "0x%x (slot %lu)", fid, msg.factory_id.size());
+        PRINT_NAMED_INFO("Robot.ConnectToBlocks.FactoryID", "0x%x (slot %lu)", fid, (unsigned long)msg.factory_id.size());
         
         if (isCharger) {
           objectsSelectedMask |= 0x80000000;
@@ -2981,7 +2981,7 @@ namespace Anki {
       _blockWorld.ClearObjectsByFamily(ObjectFamily::LightCube);
       _blockWorld.ClearObjectsByFamily(ObjectFamily::Charger);
       
-      PRINT_NAMED_INFO("Robot.ConnectToBlocks.Sending", "Num objects %lu", msg.factory_id.size());
+      PRINT_NAMED_INFO("Robot.ConnectToBlocks.Sending", "Num objects %lu", (unsigned long)msg.factory_id.size());
       return SendMessage(RobotInterface::EngineToRobot(CubeSlots(msg)));
       
     }
@@ -2995,7 +2995,7 @@ namespace Anki {
       
     Robot::ReactionCallbackIter Robot::AddReactionCallback(const Vision::Marker::Code code, ReactionCallback callback)
     {
-      //CoreTechPrint("_reactionCallbacks size = %lu\n", _reactionCallbacks.size());
+      //CoreTechPrint("_reactionCallbacks size = %lu\n", (unsigned long)_reactionCallbacks.size());
       
       _reactionCallbacks[code].emplace_front(callback);
